requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0
numpy==1.24.3
qdrant-client==1.6.2  # Fixed the typo: changed "qdrrant" to "qdrant"
torch==2.1.1 --index-url https://download.pytorch.org/whl/cpu  # Start with CPU version for compatibility
torchvision==0.16.1 --index-url https://download.pytorch.org/whl/cpu
transformers==4.36.2
#openai==1.3.5
tqdm==4.66.1
aiohttp==3.9.1
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
pandas
openai>=1.3.0
httpx>=0.25.0